diagsbackend/requirements.txt
amqp==5.0.6
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Choco==1.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-filter==2.4.0
django-macaddress==1.8.0
django-templated-mail==1.1.1
django-url-filter==0.3.15
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
enum-compat==0.0.3
gevent==21.1.2
greenlet==1.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
netaddr==0.8.0
oauthlib==3.1.1
Pillow==8.2.0
prompt-toolkit==3.0.18
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
schema==0.7.4
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.4.0
